Moved Server object
* Added new cfme.base module, with ui and ssui modules
* Moved Server object into new module
* Updated all places where Server was used
 * __init__ in appliance
